% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Albin2.R
\name{Albin2}
\alias{Albin2}
\title{Albins linreg function}
\usage{
Albin2(formula, data)
}
\arguments{
\item{formula}{shall the formula of the model}

\item{data}{shall be a matrix or data.frame}
}
\description{
This is a function simular to lm
}
