% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab4ofDOOM.R
\docType{package}
\name{Lab4ofDOOM}
\alias{Lab4ofDOOM}
\alias{Lab4ofDOOM-package}
\title{Lab4ofDOOM: A package to make a linjer regression}
\description{
This package help you to do a linjer regression with "Reference Class"
}
\section{The packages functions}{

\describe{
  \item{linreg_v2}{The main function of the package}
}
}

\author{
Albin Vasterlund and Eric Herwin
}
