% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_v2.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Do you want to make a linjer regression?}
\description{
This function make a "Reference Class" with class=linreg for a linjer regression
}
\section{Fields}{

\describe{
\item{\code{formula}}{The formula for the model}

\item{\code{data}}{A object of class data.frame}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{Give you the coef values as a numeric vector}

\item{\code{plot()}}{Printing out two good graph!}

\item{\code{pred()}}{Give you the predicted values as a numeric vector}

\item{\code{print()}}{Give you a nice view of the calculation}

\item{\code{resid()}}{Give you the residuals as a numeric vector}

\item{\code{summary()}}{Give you a nice summary of the calculation}
}}

\examples{
data(iris)
linreg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris)$print()
linreg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris)$pred()
linreg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris)$summary()
linreg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris)$resid()
linreg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris)$coef()
linreg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris)$plot()
}
