% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_class_linreg_RC.R, R/reglin.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{make_class_linreg_RC}
\alias{linreg-class}
\alias{reglin}
\title{Albins function to make RC class}
\description{
This function make a new RC class

A Reference Class to represent a bank account.
}
\section{Fields}{

\describe{
\item{\code{balance}}{A length-one numeric vector.}
}}

\section{Methods}{

\describe{
\item{\code{plot()}}{Jaja}

\item{\code{pred()}}{Muren i kina}

\item{\code{print()}}{Hej hej}

\item{\code{resid()}}{Solrosen}

\item{\code{summary()}}{tapas}

\item{\code{plot()}}{Jaja}

\item{\code{pred()}}{Muren i kina}

\item{\code{print()}}{Hej hej}

\item{\code{resid()}}{Solrosen}

\item{\code{summary()}}{tapas}
}}

