% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reglin.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{reglin}
\title{A Reference Class to represent a bank account.}
\description{
A Reference Class to represent a bank account.
}
\section{Fields}{

\describe{
\item{\code{balance}}{A length-one numeric vector.}
}}

\section{Methods}{

\describe{
\item{\code{plot()}}{Jaja}

\item{\code{pred()}}{Muren i kina}

\item{\code{print()}}{Hej hej}

\item{\code{summary()}}{tapas}
}}

